+++ client/a8c-for-agencies/sections/purchases/licenses/license-preview/index.tsx
@@ -6,8 +6,14 @@ import { getQueryArg, removeQueryArgs } from '@wordpress/url';
 import clsx from 'clsx';
 import { useTranslate } from 'i18n-calypso';
 import { useCallback, useEffect, useState, useContext, useRef } from 'react';
+import useShowFeedback from 'calypso/a8c-for-agencies/components/a4a-feedback/hooks/use-show-a4a-feedback';
+import { FeedbackType } from 'calypso/a8c-for-agencies/components/a4a-feedback/types';
 import A4APopover from 'calypso/a8c-for-agencies/components/a4a-popover';
-import { A4A_SITES_LINK_NEEDS_SETUP } from 'calypso/a8c-for-agencies/components/sidebar-menu/lib/constants';
+import {
+	A4A_SITES_LINK_NEEDS_SETUP,
+	A4A_FEEDBACK_LINK,
+	A4A_LICENSES_LINK,
+} from 'calypso/a8c-for-agencies/components/sidebar-menu/lib/constants';
 import {
 	isPressableHostingProduct,
 	isWPCOMHostingProduct,
@@ -68,6 +74,8 @@ export default function LicensePreview( {
 
 	const translate = useTranslate();
 	const dispatch = useDispatch();
+
+	const { isFeedbackShown } = useShowFeedback( FeedbackType.PurchaseCompleted );
 
 	const site = useSelector( ( state ) => getSite( state, blogId as number ) );
 	const isPressableLicense = isPressableHostingProduct( licenseKey );
@@ -282,6 +290,19 @@ export default function LicensePreview( {
 										target="_blank"
 										rel="norefferer noopener noreferrer"
 										href={ pressableManageUrl }
+										onClick={ () => {
+											if ( ! isFeedbackShown ) {
+												page.redirect(
+													addQueryArgs(
+														{
+															type: FeedbackType.PurchaseCompleted,
+															redirectUrl: A4A_LICENSES_LINK,
+														},
+														A4A_FEEDBACK_LINK
+													)
+												);
+											}
+										} }
 									>
 										{ translate( 'Manage in Pressable' ) }
 										<Icon className="gridicon" icon={ external } size={ 18 } />