+++ client/blocks/signup-form/signup-form-social-first.tsx
@@ -1,10 +1,15 @@
 import { recordTracksEvent } from '@automattic/calypso-analytics';
 import { localizeUrl } from '@automattic/i18n-utils';
 import { Button } from '@wordpress/components';
-import { useState, createInterpolateElement } from '@wordpress/element';
+import { useState, createInterpolateElement, useEffect } from '@wordpress/element';
 import { chevronLeft } from '@wordpress/icons';
 import { useI18n } from '@wordpress/react-i18n';
+import { getQueryArg } from '@wordpress/url';
 import clsx from 'clsx';
+import {
+	LOCAL_STORAGE_KEY_FOR_PG_ID,
+	LOCAL_STORAGE_KEY_FOR_PG_ID_TS,
+} from 'calypso/landing/stepper/declarative-flow/internals/steps-repository/playground/lib/initialize-playground';
 import { isGravatarOAuth2Client } from 'calypso/lib/oauth2-clients';
 import { AccountCreateReturn } from 'calypso/lib/signup/api/type';
 import { isExistingAccountError } from 'calypso/lib/signup/is-existing-account-error';
@@ -90,6 +95,17 @@ const SignupFormSocialFirst = ( {
 	const oauth2Client = useSelector( getCurrentOAuth2Client );
 	const isWoo = useSelector( getIsWoo );
 	const isGravatar = isGravatarOAuth2Client( oauth2Client );
+
+	useEffect( () => {
+		// save in localstorage for domain step in playground onboarding flow,
+		// just in case if we lose it in any possible combination of login/create-account/recover-account flow
+		const playgroundIdFromUrl = getQueryArg( window.location.href, 'playground' );
+		if ( playgroundIdFromUrl && typeof playgroundIdFromUrl === 'string' ) {
+			// ok to always overwrite with the latest playground
+			window.localStorage.setItem( LOCAL_STORAGE_KEY_FOR_PG_ID, playgroundIdFromUrl );
+			window.localStorage.setItem( LOCAL_STORAGE_KEY_FOR_PG_ID_TS, '' + Date.now() );
+		}
+	}, [] );
 
 	const renderTermsOfService = () => {
 		let tosText;