+++ client/data/php-versions/use-php-versions.jsx
@@ -1,8 +1,18 @@
 import { useTranslate } from 'i18n-calypso';
+import { useSelector } from 'calypso/state';
+import { getSelectedSiteId } from 'calypso/state/ui/selectors';
 
 export const usePhpVersions = () => {
 	const translate = useTranslate();
-	const recommendedValue = '8.2';
+	const siteId = useSelector( getSelectedSiteId );
+	// 10% of sites will have a recommended PHP version of 8.3. These sites are transferring to 8.3 by default.
+	// 243386763 is the first site of the list.
+	const isPhp83Default = siteId >= 243386763 && siteId % 10 === 0;
+	const recommendedValue = isPhp83Default ? '8.3' : '8.2';
+	const recommendedLabel = translate( '%s (recommended)', {
+		args: isPhp83Default ? '8.3' : '8.2',
+		comment: 'PHP Version for a version switcher',
+	} );
 
 	const phpVersions = [
 		{
@@ -28,14 +38,11 @@ export const usePhpVersions = () => {
 			value: '8.1',
 		},
 		{
-			label: translate( '%s (recommended)', {
-				args: '8.2',
-				comment: 'PHP Version for a version switcher',
-			} ),
-			value: recommendedValue,
+			label: isPhp83Default ? '8.2' : recommendedLabel,
+			value: '8.2',
 		},
 		{
-			label: '8.3',
+			label: isPhp83Default ? recommendedLabel : '8.3',
 			value: '8.3',
 		},
 	];