+++ client/landing/stepper/declarative-flow/flows/onboarding/onboarding.ts
@@ -1,9 +1,8 @@
-import { PLAN_PERSONAL } from '@automattic/calypso-products';
-import { OnboardSelect, Onboard, UserSelect, ProductsList } from '@automattic/data-stores';
+import { OnboardSelect } from '@automattic/data-stores';
 import { ONBOARDING_FLOW, clearStepPersistedState } from '@automattic/onboarding';
 import { useDispatch, useSelect } from '@wordpress/data';
 import { addQueryArgs, getQueryArg, getQueryArgs, removeQueryArgs } from '@wordpress/url';
-import { useState, useMemo, useEffect, useRef } from 'react';
+import { useState, useEffect } from 'react';
 import { useIsPlaygroundEligible } from 'calypso/landing/stepper/hooks/use-is-playground-eligible';
 import { SIGNUP_DOMAIN_ORIGIN } from 'calypso/lib/analytics/signup';
 import { pathToUrl } from 'calypso/lib/url';
@@ -15,29 +14,18 @@ import {
 	clearSignupCompleteFlowName,
 	clearSignupDestinationCookie,
 } from 'calypso/signup/storageUtils';
-import { useDispatch as useReduxDispatch, useSelector } from 'calypso/state';
-import { isUserLoggedIn } from 'calypso/state/current-user/selectors';
+import { useDispatch as useReduxDispatch } from 'calypso/state';
 import { setSelectedSiteId } from 'calypso/state/ui/actions';
-import {
-	STEPPER_TRACKS_EVENT_SIGNUP_START,
-	STEPPER_TRACKS_EVENT_SIGNUP_STEP_START,
-	STEPPER_TRACKS_EVENT_STEP_NAV_SUBMIT,
-} from '../../../constants';
+import { STEPPER_TRACKS_EVENT_STEP_NAV_SUBMIT } from '../../../constants';
 import { useFlowLocale } from '../../../hooks/use-flow-locale';
-import { useIsBigSkyEligible } from '../../../hooks/use-is-site-big-sky-eligible';
 import { useMarketplaceThemeProducts } from '../../../hooks/use-marketplace-theme-products';
 import { useQuery } from '../../../hooks/use-query';
-import { ONBOARD_STORE, USER_STORE } from '../../../stores';
-import { getLoginUrl } from '../../../utils/path';
+import { ONBOARD_STORE } from '../../../stores';
 import { stepsWithRequiredLogin } from '../../../utils/steps-with-required-login';
-import { useBigSkyBeforePlans } from '../../helpers/use-bigsky-before-plans-experiment';
-import { useGoalsFirstExperiment } from '../../helpers/use-goals-first-experiment';
 import { useRedirectDesignSetupOldSlug } from '../../helpers/use-redirect-design-setup-old-slug';
 import { recordStepNavigation } from '../../internals/analytics/record-step-navigation';
 import { STEPS } from '../../internals/steps';
-import { AssertConditionState, Flow, ProvidedDependencies } from '../../internals/types';
-
-const SiteIntent = Onboard.SiteIntent;
+import { Flow, ProvidedDependencies } from '../../internals/types';
 
 const clearUseMyDomainsQueryParams = ( currentStepSlug: string | undefined ) => {
 	const isDomainsStep = currentStepSlug === 'domains';
@@ -59,44 +47,7 @@ const onboarding: Flow = {
 	name: ONBOARDING_FLOW,
 	isSignupFlow: true,
 	__experimentalUseBuiltinAuth: true,
-	useTracksEventProps() {
-		const [ isLoading, isGoalsAtFrontExperiment ] = useGoalsFirstExperiment();
-		const userIsLoggedIn = useSelector( isUserLoggedIn );
-		const goals = useSelect(
-			( select ) => ( select( ONBOARD_STORE ) as OnboardSelect ).getGoals(),
-			[]
-		);
-
-		// we are only interested in the initial values and not when they values change
-		const initialGoals = useRef( goals );
-		const initialLoggedOut = useRef( ! userIsLoggedIn );
-
-		return useMemo(
-			() => ( {
-				isLoading,
-				eventsProperties: {
-					[ STEPPER_TRACKS_EVENT_SIGNUP_START ]: {
-						is_goals_first: isGoalsAtFrontExperiment.toString(),
-						...( isGoalsAtFrontExperiment && { step: 'goals' } ),
-						is_logged_out: initialLoggedOut.current.toString(),
-					},
-
-					[ STEPPER_TRACKS_EVENT_SIGNUP_STEP_START ]: {
-						...( isGoalsAtFrontExperiment && {
-							is_goals_first: 'true',
-						} ),
-						...( initialGoals.current.length && {
-							goals: initialGoals.current.join( ',' ),
-						} ),
-					},
-				},
-			} ),
-			[ isGoalsAtFrontExperiment, initialLoggedOut, initialGoals, isLoading ]
-		);
-	},
 	useSteps() {
-		// We have already checked the value has loaded in useAssertConditions
-		const [ , isGoalsAtFrontExperiment ] = useGoalsFirstExperiment();
 		const isPlaygroundEligible = useIsPlaygroundEligible();
 
 		const steps = stepsWithRequiredLogin( [
@@ -107,16 +58,6 @@ const onboarding: Flow = {
 			STEPS.PROCESSING,
 			STEPS.POST_CHECKOUT_ONBOARDING,
 		] );
-
-		if ( isGoalsAtFrontExperiment ) {
-			// Note: these steps are not wrapped in `stepsWithRequiredLogin`
-			steps.unshift(
-				STEPS.GOALS,
-				STEPS.DESIGN_CHOICES,
-				STEPS.DESIGN_SETUP,
-				STEPS.DIFM_STARTING_POINT
-			);
-		}
 
 		if ( isPlaygroundEligible ) {
 			steps.push( STEPS.PLAYGROUND );
@@ -136,27 +77,19 @@ const onboarding: Flow = {
 			setProductCartItems,
 			setSiteUrl,
 			setSignupDomainOrigin,
-			setCreateWithBigSky,
 		} = useDispatch( ONBOARD_STORE );
 		const locale = useFlowLocale();
 
-		const { planCartItem, signupDomainOrigin, isUserLoggedIn, createWithBigSky } = useSelect(
+		const { planCartItem, signupDomainOrigin } = useSelect(
 			( select ) => ( {
 				planCartItem: ( select( ONBOARD_STORE ) as OnboardSelect ).getPlanCartItem(),
 				signupDomainOrigin: ( select( ONBOARD_STORE ) as OnboardSelect ).getSignupDomainOrigin(),
-				isUserLoggedIn: ( select( USER_STORE ) as UserSelect ).isCurrentUserLoggedIn(),
-				createWithBigSky: ( select( ONBOARD_STORE ) as OnboardSelect ).getCreateWithBigSky(),
 			} ),
 			[]
 		);
 		const coupon = useQuery().get( 'coupon' );
 
 		const [ useMyDomainTracksEventProps, setUseMyDomainTracksEventProps ] = useState( {} );
-
-		const [ , isGoalsAtFrontExperiment ] = useGoalsFirstExperiment();
-		const [ , isBigSkyBeforePlansExperiment ] = useBigSkyBeforePlans();
-		const { isEligible: isBigSkyEligible } = useIsBigSkyEligible();
-		const isDesignChoicesStepEnabled = isBigSkyEligible && isGoalsAtFrontExperiment;
 
 		const { selectedMarketplaceProduct } = useMarketplaceThemeProducts();
 
@@ -167,24 +100,6 @@ const onboarding: Flow = {
 			providedDependencies: ProvidedDependencies,
 			isPlaygroundEligible: boolean
 		): [ string, string | null ] => {
-			if ( createWithBigSky && isBigSkyBeforePlansExperiment && isGoalsAtFrontExperiment ) {
-				const destination = addQueryArgs(
-					withLocale( '/setup/site-setup/launch-big-sky', locale ),
-					{
-						siteSlug: providedDependencies.siteSlug,
-						isBigSkyBeforePlansFlow: true,
-					}
-				);
-
-				return [
-					destination,
-					addQueryArgs( withLocale( '/setup/onboarding/plans', locale ), {
-						skippedCheckout: 1,
-						isBigSkyBeforePlansFlow: true,
-					} ),
-				];
-			}
-
 			if ( ! providedDependencies.hasExternalTheme && providedDependencies.hasPluginByGoal ) {
 				return [ `/home/${ providedDependencies.siteSlug }`, null ];
 			}
@@ -205,7 +120,6 @@ const onboarding: Flow = {
 
 			const destination = addQueryArgs( withLocale( '/setup/site-setup', locale ), {
 				siteSlug: providedDependencies.siteSlug,
-				...( isGoalsAtFrontExperiment && { 'goals-at-front-experiment': true } ),
 			} );
 
 			return [ destination, addQueryArgs( destination, { skippedCheckout: 1 } ) ];
@@ -216,69 +130,6 @@ const onboarding: Flow = {
 
 		const submit = async ( providedDependencies: ProvidedDependencies = {} ) => {
 			switch ( currentStepSlug ) {
-				case 'goals': {
-					const goalsUrl = withLocale( '/setup/onboarding/goals', locale );
-					const { intent } = providedDependencies;
-
-					switch ( intent ) {
-						case SiteIntent.Import: {
-							const migrationFlowLink = withLocale( '/setup/hosted-site-migration', locale );
-							return window.location.assign(
-								addQueryArgs( migrationFlowLink, {
-									back_to: goalsUrl,
-								} )
-							);
-						}
-
-						case SiteIntent.DIFM:
-							return navigate( 'difmStartingPoint' );
-
-						default: {
-							if ( isDesignChoicesStepEnabled ) {
-								return navigate( 'design-choices' );
-							}
-							return navigate( 'design-setup' );
-						}
-					}
-				}
-
-				case 'design-setup': {
-					return navigate( 'domains' );
-				}
-
-				case 'design-choices': {
-					if ( providedDependencies.destination === 'launch-big-sky' ) {
-						setCreateWithBigSky( true );
-						return navigate( 'domains' );
-					}
-					setCreateWithBigSky( false );
-					return navigate( providedDependencies.destination as string );
-				}
-
-				case 'difmStartingPoint': {
-					const difmFlowLink = addQueryArgs(
-						withLocale( '/start/website-design-services', locale ),
-						{
-							back_to: window.location.href.replace( window.location.origin, '' ),
-						}
-					);
-
-					if ( isUserLoggedIn ) {
-						return window.location.assign( difmFlowLink );
-					}
-
-					const loginUrl = getLoginUrl( {
-						variationName: flowName,
-						redirectTo: difmFlowLink,
-						locale,
-						extra: {
-							back_to: window.location.href.replace( window.location.origin, '' ),
-						},
-					} );
-
-					return window.location.assign( loginUrl );
-				}
-
 				case 'domains':
 					setSiteUrl( providedDependencies.siteUrl );
 					setDomain( providedDependencies.suggestion );
@@ -399,9 +250,6 @@ const onboarding: Flow = {
 								signup: 1,
 								checkoutBackUrl: pathToUrl( backDestination ?? '' ),
 								coupon,
-								...( createWithBigSky && isBigSkyBeforePlansExperiment && isGoalsAtFrontExperiment
-									? { [ 'big-sky-checkout' ]: 1 }
-									: {} ),
 							} )
 						);
 					} else {
@@ -418,13 +266,6 @@ const onboarding: Flow = {
 		};
 
 		return { submit };
-	},
-	useAssertConditions() {
-		const [ isLoading ] = useGoalsFirstExperiment();
-
-		return {
-			state: isLoading ? AssertConditionState.CHECKING : AssertConditionState.SUCCESS,
-		};
 	},
 	useSideEffect( currentStepSlug ) {
 		const reduxDispatch = useReduxDispatch();
@@ -445,19 +286,6 @@ const onboarding: Flow = {
 				clearSignupCompleteSlug();
 			}
 		}, [ currentStepSlug, reduxDispatch, resetOnboardStore ] );
-
-		const [ isGoalsFirstExperimentLoading, isGoalsFirstExperiment ] = useGoalsFirstExperiment();
-		// The personal plan price appears on the design choice step under these conditions. Pre-load it so it doesn't flash into existence
-		// Preload even before we know whether use is in the big-sky-before-plans experiment. By the time we know it will be too late.
-		const preloadPersonalProduct = ! isGoalsFirstExperimentLoading && isGoalsFirstExperiment;
-
-		useSelect(
-			( select ) =>
-				preloadPersonalProduct
-					? select( ProductsList.store ).getProductBySlug( PLAN_PERSONAL )
-					: undefined,
-			[ preloadPersonalProduct ]
-		);
 	},
 };