+++ client/landing/stepper/declarative-flow/internals/steps-repository/design-setup/unified-design-picker.tsx
@@ -6,10 +6,13 @@ import {
 	useDesignPickerFilters,
 } from '@automattic/design-picker';
 import { useLocale, useHasEnTranslation } from '@automattic/i18n-utils';
-import { StepContainer, ONBOARDING_FLOW, isSiteSetupFlow, Step } from '@automattic/onboarding';
+import { StepContainer, isSiteSetupFlow, Step } from '@automattic/onboarding';
 import { useSelect, useDispatch } from '@wordpress/data';
 import { useTranslate } from 'i18n-calypso';
 import { useEffect, useCallback } from 'react';
+import { useQueryProductsList } from 'calypso/components/data/query-products-list';
+import { useQuerySiteFeatures } from 'calypso/components/data/query-site-features';
+import { useQuerySitePurchases } from 'calypso/components/data/query-site-purchases';
 import { useQueryThemes } from 'calypso/components/data/query-themes';
 import FormattedHeader from 'calypso/components/formatted-header';
 import Loading from 'calypso/components/loading';
@@ -18,13 +21,11 @@ import { useActiveThemeQuery } from 'calypso/data/themes/use-active-theme-query'
 import { useIsBigSkyEligible } from 'calypso/landing/stepper/hooks/use-is-site-big-sky-eligible';
 import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
 import { useExperiment } from 'calypso/lib/explat';
-import { useDispatch as useReduxDispatch } from 'calypso/state';
 import { useActivateDesign } from '../../../../hooks/use-activate-design';
 import { useQuery } from '../../../../hooks/use-query';
 import { useSiteData } from '../../../../hooks/use-site-data';
 import { ONBOARD_STORE, SITE_STORE } from '../../../../stores';
 import { shouldUseStepContainerV2 } from '../../../helpers/should-use-step-container-v2';
-import { useGoalsFirstExperiment } from '../../../helpers/use-goals-first-experiment';
 import {
 	getDesignEventProps,
 	recordPreviewedDesign,
@@ -58,15 +59,11 @@ const UnifiedDesignPickerStep: StepType< {
 		};
 	};
 } > = ( { navigation, flow, stepName } ) => {
-	// imageOptimizationExperimentAssignment, exerimentAssignment
 	const [ isLoadingExperiment, experimentAssignment ] = useExperiment(
 		'calypso_design_picker_image_optimization_202406'
 	);
 	const variantName = experimentAssignment?.variationName;
 	const oldHighResImageLoading = ! isLoadingExperiment && variantName === 'treatment';
-
-	const [ isGoalsAtFrontExperimentLoading, isGoalsAtFrontExperiment ] = useGoalsFirstExperiment();
-	const isSiteRequired = flow !== ONBOARDING_FLOW || ! isGoalsAtFrontExperiment;
 
 	const isUpdatedBadgeDesign = useIsUpdatedBadgeDesign();
 
@@ -74,8 +71,6 @@ const UnifiedDesignPickerStep: StepType< {
 
 	const queryParams = useQuery();
 	const { goBack, submit, exitFlow } = navigation;
-
-	const reduxDispatch = useReduxDispatch();
 
 	const translate = useTranslate();
 	const locale = useLocale();
@@ -107,7 +102,6 @@ const UnifiedDesignPickerStep: StepType< {
 		)
 	);
 
-	const { setDesignOnSite, assembleSite } = useDispatch( SITE_STORE );
 	const activateDesign = useActivateDesign();
 
 	const { data: allDesigns, isLoading: isLoadingDesigns } = useStarterDesignsQuery(
@@ -213,6 +207,17 @@ const UnifiedDesignPickerStep: StepType< {
 		number: 1000,
 	} );
 
+	/**
+	 * Load data needed for the ThemeTierBadge.
+	 *
+	 * TODO: Move this within ThemeTierBadge as it's the consumer of the data.
+	 * We will need to dedupe requests because right now each ThemeTierBadge
+	 * instance will trigger a new request.
+	 */
+	useQueryProductsList();
+	useQuerySiteFeatures( [ site?.ID ] );
+	useQuerySitePurchases( site?.ID ?? -1 );
+
 	const getBadge = ( themeId: string, isLockedStyleVariation: boolean ) => (
 		<ThemeTierBadge
 			canGoToCheckout={ false }
@@ -290,31 +295,17 @@ const UnifiedDesignPickerStep: StepType< {
 					},
 					optionalProps
 				);
-			} else if ( ! isSiteRequired && ! siteSlugOrId && _selectedDesign ) {
-				handleSubmit(
-					{
-						selectedDesign: _selectedDesign,
-						selectedSiteCategory: categorization.selections?.join( ',' ),
-					},
-					optionalProps
-				);
 			}
 		},
 		[
 			activateDesign,
-			assembleSite,
-			categorization.selections,
 			designs,
 			globalStyles,
 			handleSubmit,
-			isSiteRequired,
-			reduxDispatch,
 			selectedDesign,
 			selectedStyleVariation,
-			setDesignOnSite,
 			setPendingAction,
 			setSelectedDesign,
-			site?.ID,
 			siteSlugOrId,
 		]
 	);
@@ -348,8 +339,8 @@ const UnifiedDesignPickerStep: StepType< {
 	// ********** Main render logic
 
 	// Don't render until we've done fetching all the data needed for initial render.
-	const isSiteLoading = ! site && isSiteRequired;
-	const isDesignsLoading = isLoadingDesigns || isGoalsAtFrontExperimentLoading;
+	const isSiteLoading = ! site;
+	const isDesignsLoading = isLoadingDesigns;
 	const isLoading = isSiteLoading || isDesignsLoading;
 
 	if ( isLoading || isComingFromTheUpgradeScreen ) {
@@ -441,7 +432,7 @@ const UnifiedDesignPickerStep: StepType< {
 	};
 
 	const backButton = getGoBackHandler();
-	const hideSkip = ! isGoalsAtFrontExperiment && ! isComingFromSuccessfulImport;
+	const hideSkip = ! isComingFromSuccessfulImport;
 	const skipLabelText = isComingFromSuccessfulImport
 		? translate( 'Skip to dashboard' )
 		: translate( 'Skip setup' );
@@ -449,7 +440,6 @@ const UnifiedDesignPickerStep: StepType< {
 	if ( isUsingStepContainerV2 ) {
 		return (
 			<Step.WideLayout
-				maxWidth="huge"
 				className="step-container-v2--design-picker"
 				topBar={
 					<Step.TopBar