+++ client/layout/utils.ts
@@ -1,6 +1,8 @@
+import { useRef } from 'react';
 import { shouldUseStepContainerV2 } from 'calypso/landing/stepper/declarative-flow/helpers/should-use-step-container-v2';
 import { DEFAULT_FLOW, getFlowFromURL } from 'calypso/landing/stepper/utils/get-flow-from-url';
 import { isWpMobileApp } from 'calypso/lib/mobile-app';
+
 let lastScrollPosition = 0; // Used for calculating scroll direction.
 let sidebarTop = 0; // Current sidebar top position.
 let pinnedSidebarTop = true; // We pin sidebar to the top by default.
@@ -168,10 +170,16 @@ export const handleScroll = ( event: React.UIEvent< HTMLElement > ): void => {
 	}
 };
 
-export const isRedirectingToStepContainerV2Flow = ( redirectTo: string ) => {
+const isRedirectingToStepContainerV2Flow = ( redirectTo: string ) => {
 	const { pathname, search } = new URL( redirectTo, 'http://example.com' );
 
 	return shouldUseStepContainerV2( getFlowFromURL( pathname, search ) );
+};
+
+const isMarketplaceThankYouRedirect = ( redirectTo: string ) => {
+	const { pathname, searchParams } = new URL( redirectTo, 'http://example.com' );
+
+	return pathname.startsWith( '/marketplace' ) && searchParams.has( 'onboarding' );
 };
 
 /**
@@ -199,7 +207,31 @@ export const isInStepContainerV2FlowContext = ( pathname: string, query: string 
 		if ( isRedirectingToStepContainerV2Flow( cancelTo ) ) {
 			return true;
 		}
+
+		if ( isMarketplaceThankYouRedirect( redirectTo ) ) {
+			return true;
+		}
+	}
+
+	if ( pathname.startsWith( '/marketplace' ) ) {
+		const params = new URLSearchParams( query );
+
+		if ( params.has( 'onboarding' ) ) {
+			return true;
+		}
+
+		const redirectTo = params.get( 'redirect_to' ) ?? '';
+
+		return isMarketplaceThankYouRedirect( redirectTo );
 	}
 
 	return false;
+};
+
+export const useInitialIsInStepContainerV2FlowContext = () => {
+	const ref = useRef(
+		isInStepContainerV2FlowContext( window.location.pathname, window.location.search )
+	);
+
+	return ref.current;
 };