+++ client/me/purchases/manage-purchase/index.tsx
@@ -49,6 +49,7 @@ import {
 	ProductIcon,
 	Gridicon,
 	PlanPrice,
+	MaterialIcon,
 } from '@automattic/components';
 import { Plans, type SiteDetails } from '@automattic/data-stores';
 import { localizeUrl } from '@automattic/i18n-utils';
@@ -71,6 +72,7 @@ import { bindActionCreators } from 'redux';
 import { SupportedSlugs } from 'calypso/../packages/components/src/product-icon/config';
 import googleWorkspaceIcon from 'calypso/assets/images/email-providers/google-workspace/icon.svg';
 import AsyncLoad from 'calypso/components/async-load';
+import isJetpackCrmProduct from 'calypso/components/crm-downloads/is-jetpack-crm-product';
 import QueryCanonicalTheme from 'calypso/components/data/query-canonical-theme';
 import QuerySiteDomains from 'calypso/components/data/query-site-domains';
 import QuerySitePurchases from 'calypso/components/data/query-site-purchases';
@@ -692,6 +694,36 @@ class ManagePurchase extends Component<
 		}
 
 		return null;
+	}
+
+	renderCrmDownloadsNavItem() {
+		const { purchase, translate, siteSlug } = this.props;
+
+		if ( ! purchase ) {
+			return null;
+		}
+
+		// Only show for Jetpack CRM Products
+		const productSlug = purchase.productSlug || '';
+		if ( ! isJetpackCrmProduct( productSlug ) ) {
+			return null;
+		}
+
+		const handleCrmDownloadsClick = () => {
+			recordTracksEvent( 'calypso_purchases_crm_downloads_click', {
+				product_slug: productSlug,
+			} );
+		};
+
+		// We'll pass the purchase ID in the URL, and the CRM Downloads component will fetch the actual license key
+		const path = `/purchases/crm-downloads/${ siteSlug }/${ purchase.id }`;
+
+		return (
+			<CompactCard href={ path } onClick={ handleCrmDownloadsClick }>
+				<MaterialIcon icon="person" className="card__icon" />
+				{ translate( 'CRM Downloads' ) }
+			</CompactCard>
+		);
 	}
 
 	renderRefundText() {
@@ -1418,6 +1450,7 @@ class ManagePurchase extends Component<
 						{ /* TODO: Add ability to Renew Akismet subscription */ }
 						{ ! isJetpackTemporarySitePurchase( purchase ) && this.renderUpgradeNavItem() }
 						{ this.renderEditPaymentMethodNavItem() }
+						{ config.isEnabled( 'jetpack/crm-downloads' ) && this.renderCrmDownloadsNavItem() }
 						{ this.renderReinstall() }
 						<div className="manage-purchase__downgrade-products">
 							{ config.isEnabled( 'plans/self-service-downgrade' ) && ! isPersonal( purchase )