+++ packages/data-stores/src/site/types.ts
@@ -105,6 +105,7 @@ export interface SiteDetailsPlan {
 	billing_period: string;
 	user_is_owner: boolean;
 	is_free: boolean;
+	license_key?: string;
 	features: {
 		active: string[];
 		available: Record< string, string[] >;