+++ packages/site-admin/src/index.ts
@@ -2,7 +2,15 @@
  * Components
  */
 export { Link } from './router';
-export { SidebarContent, SidebarNavigationItem, SidebarNavigationScreen } from './components/';
+
+export {
+	SidebarContent,
+	SiteHub,
+	SidebarNavigationItem,
+	SidebarNavigationScreen,
+} from './components/';
+
+export { NavigableRegion } from './interface';
 
 /**
  * Hooks